+++ examples/resources/rubrik_gcp_custom_labels/resource.tf
@@ -3,4 +3,9 @@ resource "rubrik_gcp_custom_labels" "labels" {
     "app"    = "RSC"
     "vendor" = "Rubrik"
   }
+
+  excluded_labels = [
+    "internal-cost-center",
+    "temp-*",
+  ]
 }