+++ internal/provider/framework_resource_azure_custom_tags_test.go
@@ -21,6 +21,7 @@
 package provider
 
 import (
+	"regexp"
 	"testing"
 
 	"github.com/hashicorp/terraform-plugin-testing/config"
@@ -38,11 +39,15 @@ func TestAccAzureCustomTagsResource(t *testing.T) {
 	tagKey2 := testUniqueTagKey(t)
 	tagKey3 := testUniqueTagKey(t)
 
+	exKey1 := testUniqueTagKey(t)
+	exKey2 := testUniqueTagKey(t)
+	exKey3 := testUniqueTagKey(t)
+
 	resource.Test(t, resource.TestCase{
 		ProtoV6ProviderFactories: protoV6ProviderFactories,
 		CheckDestroy:             customTagsCheckDestroy(t, core.CloudVendorAzure),
 		Steps: []resource.TestStep{{
-			// Verify that the resource can be created.
+			// Verify that the resource can be created without excluded tags.
 			Config: `
 				variable "key1" {
 					type = string
@@ -50,7 +55,6 @@ func TestAccAzureCustomTagsResource(t *testing.T) {
 				variable "key2" {
 					type = string
 				}
-
 				resource "rubrik_azure_custom_tags" "tags" {
 					custom_tags = {
 						(var.key1) = "value1"
@@ -70,14 +74,14 @@ func TestAccAzureCustomTagsResource(t *testing.T) {
 						tagKey1: knownvalue.StringExact("value1"),
 						tagKey2: knownvalue.StringExact("value2"),
 					})),
+				statecheck.ExpectKnownValue("rubrik_azure_custom_tags.tags", tfjsonpath.New(keyExcludedTags),
+					knownvalue.Null()),
 				// The override_resource_tags field defaults to true.
 				statecheck.ExpectKnownValue("rubrik_azure_custom_tags.tags", tfjsonpath.New(keyOverrideResourceTags),
 					knownvalue.Bool(true)),
 			},
 		}, {
-			// Verify that the resource can be updated. One tag is added, one
-			// is removed, one has its value changed and the override field is
-			// explicitly turned off.
+			// Verify that the resource can be created to include excluded tags.
 			Config: `
 				variable "key1" {
 					type = string
@@ -85,25 +89,92 @@ func TestAccAzureCustomTagsResource(t *testing.T) {
 				variable "key2" {
 					type = string
 				}
+				variable "exKey1" {
+					type = string
+				}
+				variable "exKey2" {
+					type = string
+				}
+				resource "rubrik_azure_custom_tags" "tags" {
+					custom_tags = {
+						(var.key1) = "value1"
+						(var.key2) = "value2"
+					}
 
+					excluded_tags = [
+						var.exKey1,
+						var.exKey2,
+					]
+				}
+			`,
+			ConfigVariables: config.Variables{
+				"key1":   config.StringVariable(tagKey1),
+				"key2":   config.StringVariable(tagKey2),
+				"exKey1": config.StringVariable(exKey1),
+				"exKey2": config.StringVariable(exKey2),
+			},
+			ConfigStateChecks: []statecheck.StateCheck{
+				statecheck.ExpectKnownValue("rubrik_azure_custom_tags.tags", tfjsonpath.New(keyID),
+					knownvalue.StringExact(azureCustomTagsID)),
+				statecheck.ExpectKnownValue("rubrik_azure_custom_tags.tags", tfjsonpath.New(keyCustomTags),
+					knownvalue.MapExact(map[string]knownvalue.Check{
+						tagKey1: knownvalue.StringExact("value1"),
+						tagKey2: knownvalue.StringExact("value2"),
+					})),
+				statecheck.ExpectKnownValue("rubrik_azure_custom_tags.tags", tfjsonpath.New(keyExcludedTags),
+					knownvalue.SetExact([]knownvalue.Check{
+						knownvalue.StringExact(exKey1),
+						knownvalue.StringExact(exKey2),
+					})),
+				// The override_resource_tags field defaults to true.
+				statecheck.ExpectKnownValue("rubrik_azure_custom_tags.tags", tfjsonpath.New(keyOverrideResourceTags),
+					knownvalue.Bool(true)),
+			},
+		}, {
+			// Verify that the resource can be updated by changing its values.
+			Config: `
+				variable "key1" {
+					type = string
+				}
+				variable "key2" {
+					type = string
+				}
+				variable "exKey1" {
+					type = string
+				}
+				variable "exKey2" {
+					type = string
+				}
 				resource "rubrik_azure_custom_tags" "tags" {
 					custom_tags = {
 						(var.key1) = "updated1"
 						(var.key2) = "value3"
 					}
 
+					excluded_tags = [
+						var.exKey1,
+						var.exKey2,
+					]
+
 					override_resource_tags = false
 				}
 			`,
 			ConfigVariables: config.Variables{
-				"key1": config.StringVariable(tagKey1),
-				"key2": config.StringVariable(tagKey3),
+				"key1":   config.StringVariable(tagKey1),
+				"key2":   config.StringVariable(tagKey3),
+				"exKey1": config.StringVariable(exKey1),
+				"exKey2": config.StringVariable(exKey3),
 			},
 			ConfigStateChecks: []statecheck.StateCheck{
 				statecheck.ExpectKnownValue("rubrik_azure_custom_tags.tags", tfjsonpath.New(keyCustomTags),
 					knownvalue.MapExact(map[string]knownvalue.Check{
 						tagKey1: knownvalue.StringExact("updated1"),
 						tagKey3: knownvalue.StringExact("value3"),
+					})),
+				statecheck.ExpectKnownValue("rubrik_azure_custom_tags.tags", tfjsonpath.New(keyExcludedTags),
+					knownvalue.SetExact([]knownvalue.Check{
+						knownvalue.StringExact(exKey1),
+						knownvalue.StringExact(exKey3),
 					})),
 				statecheck.ExpectKnownValue("rubrik_azure_custom_tags.tags", tfjsonpath.New(keyOverrideResourceTags),
 					knownvalue.Bool(false)),
@@ -119,19 +190,36 @@ func TestAccAzureCustomTagsResource(t *testing.T) {
 				variable "key2" {
 					type = string
 				}
-
+				variable "exKey1" {
+					type = string
+				}
+				variable "exKey2" {
+					type = string
+				}
 				resource "rubrik_azure_custom_tags" "tags" {
 					custom_tags = {
 						(var.key1) = "updated1"
 						(var.key2) = "value3"
 					}
+
+					excluded_tags = [
+						var.exKey1,
+						var.exKey2,
+					]
 				}
 			`,
 			ConfigVariables: config.Variables{
-				"key1": config.StringVariable(tagKey1),
-				"key2": config.StringVariable(tagKey3),
+				"key1":   config.StringVariable(tagKey1),
+				"key2":   config.StringVariable(tagKey3),
+				"exKey1": config.StringVariable(exKey1),
+				"exKey2": config.StringVariable(exKey3),
 			},
 			ConfigStateChecks: []statecheck.StateCheck{
+				statecheck.ExpectKnownValue("rubrik_azure_custom_tags.tags", tfjsonpath.New(keyExcludedTags),
+					knownvalue.SetExact([]knownvalue.Check{
+						knownvalue.StringExact(exKey1),
+						knownvalue.StringExact(exKey3),
+					})),
 				statecheck.ExpectKnownValue("rubrik_azure_custom_tags.tags", tfjsonpath.New(keyOverrideResourceTags),
 					knownvalue.Bool(true)),
 			},
@@ -145,8 +233,10 @@ func TestAccAzureCustomTagsResource(t *testing.T) {
 			ImportState:       true,
 			ImportStateVerify: true,
 			ConfigVariables: config.Variables{
-				"key1": config.StringVariable(tagKey1),
-				"key2": config.StringVariable(tagKey3),
+				"key1":   config.StringVariable(tagKey1),
+				"key2":   config.StringVariable(tagKey3),
+				"exKey1": config.StringVariable(exKey1),
+				"exKey2": config.StringVariable(exKey3),
 			},
 		}, {
 			// import {} block with id attribute. Note, the import will take
@@ -158,9 +248,132 @@ func TestAccAzureCustomTagsResource(t *testing.T) {
 			ImportStateId:   "dummy",
 			ImportState:     true,
 			ConfigVariables: config.Variables{
-				"key1": config.StringVariable(tagKey1),
-				"key2": config.StringVariable(tagKey3),
+				"key1":   config.StringVariable(tagKey1),
+				"key2":   config.StringVariable(tagKey3),
+				"exKey1": config.StringVariable(exKey1),
+				"exKey2": config.StringVariable(exKey3),
 			},
+		}},
+	})
+}
+
+// TestAccAzureCustomTagsResource_ExcludedTagsOnly verifies that a resource can
+// manage excluded tags without any custom tags.
+func TestAccAzureCustomTagsResource_ExcludedTagsOnly(t *testing.T) {
+	exKey1 := testUniqueTagKey(t)
+	exKey2 := testUniqueTagKey(t)
+	exKey3 := testUniqueTagKey(t)
+
+	resource.Test(t, resource.TestCase{
+		ProtoV6ProviderFactories: protoV6ProviderFactories,
+		CheckDestroy:             customTagsCheckDestroy(t, core.CloudVendorAzure),
+		Steps: []resource.TestStep{{
+			// Verify that the resource can be created with only excluded tags.
+			Config: `
+				variable "exKey1" {
+					type = string
+				}
+				variable "exKey2" {
+					type = string
+				}
+				resource "rubrik_azure_custom_tags" "tags" {
+					excluded_tags = [
+						var.exKey1,
+						var.exKey2,
+					]
+				}
+			`,
+			ConfigVariables: config.Variables{
+				"exKey1": config.StringVariable(exKey1),
+				"exKey2": config.StringVariable(exKey2),
+			},
+			ConfigStateChecks: []statecheck.StateCheck{
+				statecheck.ExpectKnownValue("rubrik_azure_custom_tags.tags", tfjsonpath.New(keyID),
+					knownvalue.StringExact(azureCustomTagsID)),
+				statecheck.ExpectKnownValue("rubrik_azure_custom_tags.tags", tfjsonpath.New(keyCustomTags),
+					knownvalue.Null()),
+				statecheck.ExpectKnownValue("rubrik_azure_custom_tags.tags", tfjsonpath.New(keyExcludedTags),
+					knownvalue.SetExact([]knownvalue.Check{
+						knownvalue.StringExact(exKey1),
+						knownvalue.StringExact(exKey2),
+					})),
+				statecheck.ExpectKnownValue("rubrik_azure_custom_tags.tags", tfjsonpath.New(keyOverrideResourceTags),
+					knownvalue.Bool(true)),
+			},
+		}, {
+			// Verify that the resource can be updated. One excluded tag is
+			// added and one is removed, still without any custom tags.
+			Config: `
+				variable "exKey1" {
+					type = string
+				}
+				variable "exKey2" {
+					type = string
+				}
+				resource "rubrik_azure_custom_tags" "tags" {
+					excluded_tags = [
+						var.exKey1,
+						var.exKey2,
+					]
+				}
+			`,
+			ConfigVariables: config.Variables{
+				"exKey1": config.StringVariable(exKey1),
+				"exKey2": config.StringVariable(exKey3),
+			},
+			ConfigStateChecks: []statecheck.StateCheck{
+				statecheck.ExpectKnownValue("rubrik_azure_custom_tags.tags", tfjsonpath.New(keyCustomTags),
+					knownvalue.Null()),
+				statecheck.ExpectKnownValue("rubrik_azure_custom_tags.tags", tfjsonpath.New(keyExcludedTags),
+					knownvalue.SetExact([]knownvalue.Check{
+						knownvalue.StringExact(exKey1),
+						knownvalue.StringExact(exKey3),
+					})),
+			},
+		}},
+	})
+}
+
+// TestAccAzureCustomTagsResource_InvalidConfigs verifies that the custom_tags and
+// excluded_tags fields reject an explicitly empty collection, and that at least
+// one of them must be specified. Terraform distinguishes an unset field from an
+// empty one, and only the unset form is meaningful for these fields.
+func TestAccAzureCustomTagsResource_InvalidConfigs(t *testing.T) {
+	tagKey := testUniqueTagKey(t)
+
+	resource.Test(t, resource.TestCase{
+		ProtoV6ProviderFactories: protoV6ProviderFactories,
+		Steps: []resource.TestStep{{
+			Config: `
+				resource "rubrik_azure_custom_tags" "tags" {
+					custom_tags = {}
+				}
+			`,
+			ExpectError: regexp.MustCompile(`custom_tags map must contain at least 1 elements`),
+		}, {
+			Config: `
+				variable "key" {
+					type = string
+				}
+
+				resource "rubrik_azure_custom_tags" "tags" {
+					custom_tags = {
+						(var.key) = "value"
+					}
+
+					excluded_tags = []
+				}
+			`,
+			ConfigVariables: config.Variables{
+				"key": config.StringVariable(tagKey),
+			},
+			ExpectError: regexp.MustCompile(`excluded_tags set must contain at least 1 elements`),
+		}, {
+			Config: `
+				resource "rubrik_azure_custom_tags" "tags" {
+				}
+			`,
+			ExpectError: regexp.MustCompile(`At least one of these attributes must be configured:\s+\[custom_tags,excluded_tags\]`),
 		}},
 	})
 }